Add new game mechanics. Game script completed.
Added a game function, calls the playRound function 5 times. Keeps score of each round. After 5 rounds, the winner is declared and the game is over.

playRound: Function logs to the console the winner and loser of the round, and returns to the game function which player won the current round.